+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/action/ActionExecutor.scala
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-package cloudflow.operator
-package action
+package cloudflow.operator.action
 
 import scala.concurrent.Future
