+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/action/Actions.scala
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-package cloudflow.operator
-package action
+package cloudflow.operator.action
 
 import scala.collection.immutable._
 import skuber._
