+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/action/EventActions.scala
@@ -20,7 +20,6 @@ import java.time.ZonedDateTime
 import cloudflow.blueprint.deployment.StreamletDeployment
 import cloudflow.operator.action.EventActions.EventType.EventType
 import cloudflow.operator.action.runner.Runner
-import cloudflow.operator.{ CloudflowApplication, CloudflowLabels }
 import skuber.json.format.eventFmt
 import skuber.{ Event, ObjectEditor, ObjectMeta, ObjectResource }
