+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/action/Name.scala
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package cloudflow.operator
+package cloudflow.operator.action
 
 import java.text.Normalizer
