+++ core/cloudflow-operator-actions/src/main/scala/cloudflow/operator/action/runner/FlinkRunner.scala
@@ -32,7 +32,7 @@ import skuber._
 import cloudflow.blueprint.deployment._
 import FlinkResource._
 
-import cloudflow.operator._
+import cloudflow.operator.action._
 import cloudflow.operator.action.Action
 
 object FlinkRunner {
