+++ core/cloudflow-operator-actions/src/test/scala/cloudflow/operator/action/EndpointActionsSpec.scala
@@ -14,8 +14,7 @@
  * limitations under the License.
  */
 
-package cloudflow.operator
-package action
+package cloudflow.operator.action
 
 import org.scalatest._
