+++ core/cloudflow-operator-actions/src/test/scala/cloudflow/operator/action/EventActionsSpec.scala
@@ -19,7 +19,6 @@ package cloudflow.operator.action
 import org.scalatest._
 import cloudflow.blueprint._
 import BlueprintBuilder._
-import cloudflow.operator.{ CloudflowApplication, CloudflowApplicationSpecBuilder, TestDeploymentContext }
 import cloudflow.operator.action.runner.AkkaRunner
 
 class EventActionsSpec extends WordSpec with MustMatchers with GivenWhenThen with EitherValues with Inspectors with TestDeploymentContext {
