+++ core/cloudflow-operator-actions/src/test/scala/cloudflow/operator/action/TopicActionsSpec.scala
@@ -14,9 +14,7 @@
  * limitations under the License.
  */
 
-package cloudflow.operator
-package action
-
+package cloudflow.operator.action
 import org.scalatest._
 import cloudflow.blueprint.{ Topic => BTopic, _ }
 import BlueprintBuilder._
