+++ core/cloudflow-operator-actions/src/test/scala/cloudflow/operator/action/runner/AkkaRunnerSpec.scala
@@ -18,10 +18,11 @@ package cloudflow.operator.action.runner
 
 import cloudflow.blueprint._
 import cloudflow.blueprint.deployment.{ PrometheusConfig, StreamletDeployment }
-import cloudflow.operator.{ CloudflowApplication, CloudflowApplicationSpecBuilder, TestDeploymentContext }
 import com.typesafe.config.ConfigFactory
 import org.scalatest._
 import skuber.{ Volume, _ }
+
+import cloudflow.operator.action._
 
 class AkkaRunnerSpec extends WordSpecLike with OptionValues with MustMatchers with GivenWhenThen with TestDeploymentContext {
