+++ core/cloudflow-operator-actions/src/test/scala/cloudflow/operator/action/runner/SparkRunnerSpec.scala
@@ -23,8 +23,8 @@ import skuber._
 
 import cloudflow.blueprint._
 import cloudflow.blueprint.deployment.{ PrometheusConfig, StreamletDeployment }
+import cloudflow.operator.action._
 import cloudflow.operator.action.runner.SparkResource.{ AlwaysRestartPolicy, CR }
-import cloudflow.operator._
 
 class SparkRunnerSpec extends WordSpecLike with OptionValues with MustMatchers with GivenWhenThen with TestDeploymentContext {
